+++ activities/Tangram.activity/js/Result.js
@@ -132,7 +132,7 @@ var Result = {
       },
       actionButtons: {
         width: 30,
-        width: 30,
+        height: 30,
       },
       totalScore: 0,
       puzzlesSet: [],