+++ CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Curriculum Activity: blank white screen due to TypeError when entering achievements page #2073
 - PhysicsJS newton/gravity settings are not saved after leaving #2072
 - Chess Activity: AI crashes due to undefined variables (missing 'this' and 'flags') #2062
+- Duplicate 'width' property in actionButtons across Tangram activity #2090
 
 ## [1.9.0] - 2025-03-25
 ### Added